[2505-CHORE-377] rename cover-confirm → cover/confirm route, update caller
